requirements.txt
discord.py
python-dotenv
aioconsole
discord.py[voice]